# gh-97982: Reuse `PyUnicode_Count` in `unicode_count`