--- a/Misc/NEWS.d/next/Core and Builtins/2022-10-07-13-09-57.gh-issue-97982.6qR7Qm.rst
+++ b/Misc/NEWS.d/next/Core and Builtins/2022-10-07-13-09-57.gh-issue-97982.6qR7Qm.rst
@@ -1,2 +1,3 @@
-Reuse ``PyUnicode_Count`` function in ``unicode_count`` in
+Create ``any_unicode_count`` private helper for 
+both ``PyUnicode_Count`` and ``unicode_count`` in
 ``unicodeobject.c``.