--- a/Misc/NEWS.d/next/Core and Builtins/2022-10-07-13-09-57.gh-issue-97982.6qR7Qm.rst
+++ b/Misc/NEWS.d/next/Core and Builtins/2022-10-07-13-09-57.gh-issue-97982.6qR7Qm.rst
@@ -1,3 +1,0 @@
-Create ``any_unicode_count`` private helper for
-both ``PyUnicode_Count`` and ``unicode_count`` in
-``unicodeobject.c``.