This now makes `unicode_count` as optimized as `PyUnicode_Count`. There was a missing optimization, see https://github.com/python/cpython/commit/c3cec7868bf1019c0987f1e9aadb56d73fa93d61

However, I've decided not to touch `anylib_count`, because of [`(const Py_UCS1*)buf1` (and similar) casts](https://github.com/python/cpython/blob/cbdeda8ce7a3543cb3376d70e4cd46fcf24f42a7/Objects/unicodeobject.c#L9005) and function ordering. We cannot directly call `anylib_count` from `PyUnicode_Count`. I don't think it is worth creating forward declarations / moving things just for that.



<!-- gh-issue-number: gh-97982 -->
* Issue: gh-97982
<!-- /gh-issue-number -->