@vstinner 2022-10-07
cc @serhiy-storchaka