@vstinner 2022-10-10
Merging two functions which are basically the same is a good thing.

> This now makes unicode_count as optimized as PyUnicode_Count. There was a missing optimization, see https://github.com/python/cpython/commit/c3cec7868bf1019c0987f1e9aadb56d73fa93d61

But I'm not convinced that asciilib_count() is different than ucs1lib_count(). It seems like the code is exactly the same, it doesn't see to use the property that the string is ASCII only.

> » pyperf timeit '"abcabd".count("ab")'

* before: Mean +- std dev: 495 ns +- 7 ns
* after: Mean +- std dev: 459 ns +- 9 ns

I'm not convinced. You didn't explain how you built Python. Did you enable LTO and/or PGO? Which is your OS? Which compiler?