@vstinner 2022-10-10
> I ran a benchmark: there is no difference. I would suggest to remove asciilib_count().

I wrote PR #98164 for that.