@vstinner 2022-10-11
> If asciilib_count and ucs1lib_count are the same, why there is a difference in https://github.com/python/cpython/pull/98025#issuecomment-1272275910?

These timings are about 500 ns, whereas I got timings about 60 ns. Maybe the benchmark with 500 ns was run on Python built in debug mode? As I wrote, I would like to know more about this benchmark: OS name and version, compiler name and options, how was Python built (debug mode or not? etc.), ...