@vstinner 2022-10-11
I remove asciilib_count() function in commit https://github.com/python/cpython/commit/df3a6d9beb8a7a3fe87a6d4126384fd3e0213853.